+++ tests/MeuValorLiquido.WebApp.Tests/PublicPagesTests.cs
@@ -6,10 +6,12 @@ namespace MeuValorLiquido.WebApp.Tests;
 
 public class PublicPagesTests : IClassFixture<WebApplicationFactory<Program>>
 {
+    private readonly WebApplicationFactory<Program> factory;
     private readonly HttpClient client;
 
     public PublicPagesTests(WebApplicationFactory<Program> factory)
     {
+        this.factory = factory;
         client = factory.WithWebHostBuilder(builder => builder.UseEnvironment("Testing")).CreateClient();
     }
 
@@ -17,6 +19,7 @@ public class PublicPagesTests : IClassFixture<WebApplicationFactory<Program>>
     [InlineData("/")]
     [InlineData("/calculadoras")]
     [InlineData("/calculadoras/salario-liquido")]
+    [InlineData("/calculadoras/salario-bruto-necessario")]
     [InlineData("/sobre")]
     [InlineData("/contato")]
     [InlineData("/politica-de-privacidade")]
@@ -31,5 +34,18 @@ public class PublicPagesTests : IClassFixture<WebApplicationFactory<Program>>
         var response = await client.GetAsync(url);
 
         response.IsSuccessStatusCode.Should().BeTrue();
+    }
+
+    [Fact]
+    public async Task Salary_Bruto_Alias_Should_Redirect_To_Calculator()
+    {
+        using var noRedirectClient = factory
+            .WithWebHostBuilder(builder => builder.UseEnvironment("Testing"))
+            .CreateClient(new WebApplicationFactoryClientOptions { AllowAutoRedirect = false });
+
+        var response = await noRedirectClient.GetAsync("/calculadora-salario-bruto");
+
+        response.StatusCode.Should().Be(System.Net.HttpStatusCode.MovedPermanently);
+        response.Headers.Location!.ToString().Should().Contain("salario-bruto-necessario");
     }
 }